+++ packages/tailwindcss/src/canonicalize-candidates.ts
@@ -1796,6 +1796,28 @@ function modernizeArbitraryValuesVariant(
         continue
       }
 
+      // `[&:has(…)]` can be replaced with `has-[…]`
+      if (
+        // Only top-level, so `group-[&:has(…)]` is not covered
+
+        parent === null &&
+        // [&:has(…)]:flex
+        //  ^ ^^^^^^
+        ast.length === 2 &&
+        ast[0].kind === 'selector' &&
+        ast[0].value === '&' &&
+        ast[1].kind === 'function' &&
+        ast[1].value === ':has' &&
+        ast[1].nodes.length === 1 &&
+        ast[1].nodes[0].kind === 'selector'
+      ) {
+        replaceObject(
+          variant,
+          designSystem.parseVariant(`has-[${SelectorParser.toCss(ast[1].nodes)}]`),
+        )
+        continue
+      }
+
       // `in-*` variant. If the selector ends with ` &`, we can convert it to an
       // `in-*` variant.
       //
@@ -2117,7 +2139,7 @@ function optimizeArbitraryValueExpressions(
 
   // Start by constant folding the value expression, when dealing with `calc(…)`
   if (valueAst.length === 1 && valueAst[0].kind === 'function' && valueAst[0].value === 'calc') {
-    let [folded, foldedValueAst] = constantFoldDeclarationAst(valueAst)
+    let [folded, foldedValueAst] = constantFoldDeclarationAst(valueAst, null, false)
     if (folded) {
       let replacement = cloneCandidate(candidate)
       replacement.value!.value = ValueParser.toCss(foldedValueAst)
@@ -2139,7 +2161,7 @@ function optimizeArbitraryValueExpressions(
     // Move `* -1` inside, and try to constant fold to see if it's even worth
     // updating the candidate or not.
     let expressionAst = ValueParser.parse(`calc(${candidate.value!.value} * -1)`)
-    let [folded, foldedExpressionAst] = constantFoldDeclarationAst(expressionAst)
+    let [folded, foldedExpressionAst] = constantFoldDeclarationAst(expressionAst, null, false)
     if (folded) {
       let replacement = cloneCandidate(candidate)